+++ examples/tencentcloud-image/variables.tf
@@ -1,0 +1,3 @@
+variable "image_imstance_name" {
+    default = "image-instance-keep"
+}