+++ tencentcloud/extension_images.go
@@ -1,0 +1,6 @@
+package tencentcloud
+
+const (
+	TRUE  = "true"
+	FALSE = "false"
+)